A path search is sliced across ticks now, not truncated
The original has no per-search cap at all. It slices its A* at a hundred
expansions a tick (0x40EEAF) and resumes the same search on the next one,
so a long route completes over several ticks and nothing is ever
truncated. RWE capped at a thousand and handed back a partial path the unit
walked and then re-requested from -- the one structural piece of that
scheduler with no equivalent here, and what the cap existed in place of.

AStarPathFinder is a state machine now: beginSearch, stepSearch(n),
takeResult. PathFindingService gives a search whatever is left of the
tick's budget and keeps it exactly where it stopped if that runs out.
There is no per-search cap left at all -- a search ends at the goal or at
an empty open list -- so a partial result means the place genuinely cannot
be reached and nothing else, and the branch that used to aim at the goal
anyway is gone with the cap that produced it.

One search runs at a time, because the scratch grid stamps its cells with
the search they belong to and starting a second would stamp the first one's
cells stale. The original works under the same constraint from the other
end: its pathfinder is a singleton at gm+0x14207 with one set of map
entries, which is why it can slice at all. A suspended search is re-checked
against its unit before it is resumed, and dropped if the unit died or was
given a different order.

A save carries it, in five integers. A half-finished A* cannot be
serialised and does not have to be: a search is a pure function of the
world, the goal and the cell it started from. The first two are in the save
already; only the third is lost, because a unit whose search is suspended
keeps walking its straight-line stand-in and is no longer standing there.
So the save writes that footprint and the expansion count, and the load
rebuilds the search and runs it forward that far -- exactly, A* being
deterministic. Dropping the search on both sides instead would have passed
the saved-game test and broken replays outright, whose keyframes go through
the same saveSimulationToJson during playback against a recording that
dropped nothing.

The scratch cell went from eight bytes to twelve. The static asserts that
made its int16 indices safe were bounded by the cap that is gone, and a
search that can close every cell of a 512x512 map needs more than an int16
addresses. The budget stays at 4000 a tick, forty times the original's
hundred: raising it is its own roadmap entry, and moving both at once would
make neither measurable.

Three tests. A search cut into one-expansion slices returns the route an
uninterrupted search returns; a goal walled off on a 64x64 map comes back
exhausted after four thousand expansions where it used to stop at a
thousand with the open list still full; and a game saved mid-search replays
tick for tick against the one loaded back from it.

And path_bench's obstacles were broken all along -- an immobile unit
definition with no yard map, an assert in Debug and an empty-optional
dereference in Release -- so every figure it has produced was measured with
obstacles that did not reliably obstruct. It has a yard map now, and the
numbers the roadmap quoted from it are withdrawn rather than restated,
because the comparison has not been re-run.

+++ src/rwe/pathfinding/AStarScratch.h
@@ -33,12 +33,20 @@ namespace rwe
         static constexpr uint8_t WaterValue = 1u << 5;
 
         /**
-         * Eight bytes, and every field the search wants about a cell is in
-         * them. Worth the narrow types: the open list of a thousand-vertex
-         * search spans a few thousand cells, and a sift step reaches into one
-         * of them at random, so whether the set fits in L1 decides what a sift
-         * costs. The indices are narrow safely -- see the static asserts in
-         * AStarPathFinder, which bound both against the search's own limits.
+         * Twelve bytes, and every field the search wants about a cell is in
+         * them. The narrow stamp and flags are worth keeping: a sift step
+         * reaches into one of these at random, so whether the working set
+         * fits in L1 decides what a sift costs.
+         *
+         * The two indices used to be int16, bounded by a static assert
+         * against a per-search cap of a thousand expansions. That cap is
+         * gone -- a search now runs to completion over as many ticks as it
+         * takes, as the original's does -- and with it the bound. A search
+         * that expands every cell of a 512x512 map closes 262144 vertices and
+         * can hold a frontier far past what an int16 addresses, so both
+         * indices are int32 and the asserts they justified have gone with
+         * them. Four bytes a cell, and the map's grid is the only thing that
+         * bounds a search now.
          */
         struct Cell
         {
@@ -48,9 +56,9 @@ namespace rwe
             uint8_t flags{0};
             uint8_t unused{0};
             /** Position in the open heap, or -1 when this cell is not open. */
-            int16_t openIndex{-1};
+            int32_t openIndex{-1};
             /** Position in the closed list, or -1 when this cell is not closed. */
-            int16_t closedIndex{-1};
+            int32_t closedIndex{-1};
         };
 
         std::vector<Cell> cells;
